aurora/backend/requirements.txt
# Aurora Backend Dependencies

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Supabase
supabase>=2.3.0

# AI
groq>=0.4.0
sentence-transformers>=3.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=9.0.0
cachetools>=5.5.0
python-jose[cryptography]>=3.3.0
tiktoken>=0.5.0
apscheduler>=3.10.0
Pillow>=10.0.0
python-multipart>=0.0.6
